[ui] Move UI constants into UI folder
